disallow (mutual) recursive pattern binding definition
It is not clear how to type check the Haskell Core code that is
generated from such definitions. Since definitions of this form are
rather niche, it is better to disallow them altogether (even though
the type checking of the generated Haskell Core code is purely for the
purpose of debugging and sanity checking the plugin implementation).