annotate tests to anticipate warnings and errors
I've added a new annotation type InternalAnn to be used for testing
the typechecker (it is not intended for regular use). If a function is
annotated with ExpectError or ExpectWarning, the type checker will
ignore any error or warning, respectively. On the other hand, the
typechecker will produce an error messages if a function is annotated
with ExpectError or ExpectWarning but no such error or warning is
produced.